- Make changes to use refactored mobile UI
